+++ TradingSystem/Data/TransactionData.cs
@@ -7,8 +7,8 @@
         public Guid SellerId { get; set; }
         public string InstrumentId { get; set; }
         public int Size { get; set; }
-        public float Price { get; set; }
-        public float SpreadPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal SpreadPrice { get; set; }
         public DateTime Time { get; set; } //Maybe string is better? depends on implementation i guess. Left as DateTime object for now
         public bool Succeeded { get; set; }
