+++ TradingSystem/Logic/ExternalBrokers/NASDAQAPI.cs
@@ -7,7 +7,7 @@ namespace TradingSystem.Logic.ExternalBrokers
     public interface INASDAQ
     {
         public Stock simulatePriceChange(int simSpeed, ref Lock simulationLock, ref CancellationToken token, ref bool first);
-        public Dictionary<string, float> getPrices();
+        public Dictionary<string, decimal> getPrices();
         public List<string> getStocks();
     }
 
@@ -15,7 +15,7 @@ namespace TradingSystem.Logic.ExternalBrokers
     {
         private readonly ILogger<NASDAQAPI> _logger;
         private List<string> myStocks = new();
-        private Dictionary<string, float> myPrices = new Dictionary<string, float>();
+        private Dictionary<string, decimal> myPrices = new Dictionary<string, decimal>();
         private Random rand = new Random();
         
         public NASDAQAPI(ILogger<NASDAQAPI> logger, IOptions<BrokerStocks> brokerStocks)
@@ -24,12 +24,12 @@ namespace TradingSystem.Logic.ExternalBrokers
             var options = brokerStocks.Value;
             foreach (string name in options.NASDAQ)
             {
-                myPrices.Add(name, 10.0f);
+                myPrices.Add(name, 10.0m);
                 myStocks.Add(name);
             }
         }
 
-        public Dictionary<string, float> getPrices()
+        public Dictionary<string, decimal> getPrices()
         {
             return myPrices;
         }
@@ -55,7 +55,7 @@ namespace TradingSystem.Logic.ExternalBrokers
                 {
                     first = false;
                     var updateKey = myPrices.ElementAt(rand.Next(0, myPrices.Count)).Key;
-                    var price = (rand.Next(0, 2) > 0) ? myPrices[updateKey] - 0.1f : myPrices[updateKey] + 0.1f;
+                    var price = (rand.Next(0, 2) > 0) ? myPrices[updateKey] - 0.1m : myPrices[updateKey] + 0.1m;
                     _logger.NasdaqApiUpdatePrice(updateKey, myPrices[updateKey], price);
                     myPrices[updateKey] = price;
                     var updatedStock = new Stock
